Add a way to remove/add dom events to the Nel class.
Closed #16 Improved handling of DOM events.

A list of pairs of event names and event handlers has been added to the Nel class. The handlers listed in the list are removed from the event listeners just before the Nel element is builded or unmounted, and are added to the event listeners just after the builded element is mounted.

On branch develop
Changes to be committed:
	modified:   core/fr.js