Improve to pass ts-check in core/fr.js
I thought ts-check was enabled, but it turns out I had entered the
keywords incorrectly, so ts-check was disabled. When I enabled it, I
got a bunch of error messages, so I organized the code and made some
improvements.
And add comment to leave legal comment on esbuild.
- Major changes (improvements)
  - Moved global variables under window.__proto__ to under globalThis
  - Changed RefKey from a commented typedef to an actual class
	definition
	- Deleted getter / setter from NodeElement interface class
	- The Nel class has deprecated NType.tempLiteral in its build method

On branch develop
Changes to be committed:
	modified:   core/fr.js