Fix JSDOC error removal
When compiling with jsdoc, an error occurred. This was because the jsdoc comment contained a "ts" expression. The "ts" expression was written to silence //@ts-check, but I prioritized compiling jsdoc and removed the "ts" expression. Instead, I inserted //@ts-ignore to silence //@ts-check. This is a bit off-topic, but I compromised as a trade-off.
And add docdash for jsdoc documentation.

On branch develop
Changes to be committed:
	modified:   core/fr.js
	modified:   core/reactive.js
	modified:   jsdoc.json
	modified:   package-lock.json
	modified:   package.json.in